refactor: 메서드명 변경
